+++ lib/src/deriv_chart/chart/chart.dart
@@ -2,7 +2,6 @@ import 'package:collection/collection.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/models/chart_scale_model.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/mobile_chart_frame_dividers.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/x_axis/x_axis_model.dart';
-import 'package:deriv_chart/src/deriv_chart/interactive_layer/interactive_layer_base.dart';
 import 'package:deriv_chart/src/theme/dimens.dart';
 import 'package:flutter/foundation.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/gestures/gesture_manager.dart';
@@ -22,6 +21,7 @@ import '../../misc/chart_controller.dart';
 import '../../models/tick.dart';
 import '../../theme/chart_default_dark_theme.dart';
 import '../../theme/chart_theme.dart';
+import '../interactive_layer/interactive_layer_behaviours/interactive_layer_behaviour.dart';
 import 'bottom_chart.dart';
 import 'bottom_chart_mobile.dart';
 import 'data_visualization/annotations/chart_annotation.dart';
