+++ lib/src/deriv_chart/deriv_chart.dart
@@ -25,6 +25,10 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
+import 'interactive_layer/interactive_layer_behaviours/interactive_layer_behaviour.dart';
+import 'interactive_layer/interactive_layer_behaviours/interactive_layer_desktop_behaviour.dart';
+import 'interactive_layer/interactive_layer_behaviours/interactive_layer_mobile_behaviour.dart';
+
 /// A wrapper around the [Chart] which handles adding indicators to the chart.
 class DerivChart extends StatefulWidget {
   /// Initializes
