+++ lib/src/deriv_chart/interactive_layer/interactable_drawings/trend_line/line_interactable_drawing.dart
@@ -11,7 +11,8 @@ import 'package:deriv_chart/src/theme/painting_styles/line_style.dart';
 import 'package:flutter/widgets.dart';
 
 import '../../helpers/paint_helpers.dart';
-import '../../interactive_layer_base.dart';
+import '../../interactive_layer_behaviours/interactive_layer_desktop_behaviour.dart';
+import '../../interactive_layer_behaviours/interactive_layer_mobile_behaviour.dart';
 import '../drawing_adding_preview.dart';
 import '../drawing_v2.dart';
 import '../interactable_drawing.dart';
